refactor(domain): split value_objects.py into package
Convert single value_objects.py file into a structured package:
- identifiers.py: RunId, StageName, EntityName, PipelineId, ChemblId
- crypto.py: HashDigest
- network.py: HttpUrl
- temporal.py: Timestamp
- __init__.py: re-exports all for backward compatibility

All imports remain backward compatible:
  from bioetl.domain.value_objects import ChemblId
